Initial Dbus service and client (#53)
Includes a base implementation for dbus, based on what's in plusdeck. In addition, it includes a domain mapping package in `crystalfontz.dbus.domain`.

It does not include tests, systemd files, reporting, effects, special characters or statuses, and a LOT of follow-up work needs to be done. But it *has* been ticketed.